refactor(fe): refactor public directory (#2188)
* refactor(re): refactor public folde

* Optimised images with calibre/image-actions

* fix(fe): fix directory

* fix(fe): fix file name

---------
