+++ apps/frontend/app/(client)/(main)/contest/[contestId]/layout.tsx
@@ -1,9 +1,9 @@
 import ContestStatusTimeDiff from '@/components/ContestStatusTimeDiff'
+import { fetcher, fetcherWithAuth, getStatusWithStartEnd } from '@/lib/utils'
 import { auth } from '@/libs/auth'
-import { fetcher, fetcherWithAuth, getStatusWithStartEnd } from '@/libs/utils'
 import { dateFormatter } from '@/libs/utils'
-import calendarIcon from '@/public/icons/calendar.svg'
-import checkIcon from '@/public/icons/check-blue.svg'
+import Calendar from '@/public/20_calendar.svg'
+import CheckIcon from '@/public/check_blue.svg'
 import type { Contest } from '@/types/type'
 import Image from 'next/image'
 import ContestTabs from '../_components/ContestTabs'
