+++ apps/frontend/app/(client)/(main)/settings/_components/NewPwSection.tsx
@@ -1,7 +1,7 @@
 import { Input } from '@/components/ui/input'
 import { cn } from '@/libs/utils'
-import invisibleIcon from '@/public/icons/invisible.svg'
-import visibleIcon from '@/public/icons/visible.svg'
+import invisible from '@/public/24_invisible.svg'
+import visible from '@/public/24_visible.svg'
 import type { SettingsFormat } from '@/types/type'
 import Image from 'next/image'
 import React from 'react'
