+++ apps/frontend/app/admin/contest/[contestId]/(overall)/layout.tsx
@@ -4,7 +4,7 @@ import KatexContent from '@/components/KatexContent'
 import { Button } from '@/components/ui/button'
 import { GET_CONTEST } from '@/graphql/contest/queries'
 import { dateFormatter } from '@/libs/utils'
-import periodIcon from '@/public/icons/period.svg'
+import Period from '@/public/period.svg'
 import { useQuery } from '@apollo/client'
 import Image from 'next/image'
 import Link from 'next/link'
