+++ apps/frontend/app/admin/contest/_components/ContestTableColumns.tsx
@@ -11,8 +11,8 @@ import {
 } from '@/components/ui/tooltip'
 import { UPDATE_CONTEST_VISIBLE } from '@/graphql/contest/mutations'
 import { cn, dateFormatter } from '@/libs/utils'
-import invisibleIcon from '@/public/icons/invisible.svg'
-import visibleIcon from '@/public/icons/visible.svg'
+import InvisibleIcon from '@/public/24_invisible.svg'
+import VisibleIcon from '@/public/24_visible.svg'
 import { useMutation } from '@apollo/client'
 import * as TooltipPrimitive from '@radix-ui/react-tooltip'
 import type { ColumnDef, Row } from '@tanstack/react-table'
