+++ apps/frontend/components/EditorDescription.tsx
@@ -16,10 +16,10 @@ import {
   DialogTrigger
 } from '@/components/ui/dialog'
 import { convertToLetter } from '@/libs/utils'
-import compileIcon from '@/public/icons/compile-version.svg'
-import copyBlueIcon from '@/public/icons/copy-blue.svg'
-import copyCompleteIcon from '@/public/icons/copy-complete.svg'
-import copyIcon from '@/public/icons/copy.svg'
+import CopyIcon from '@/public/24_copy.svg'
+import compileIcon from '@/public/compileVersion.svg'
+import copyIcon from '@/public/copy.svg'
+import copyCompleteIcon from '@/public/copyComplete.svg'
 import type { ContestProblem, ProblemDetail } from '@/types/type'
 import type { Level } from '@/types/type'
 import { motion } from 'framer-motion'
