+++ apps/frontend/components/EditorLayout.tsx
@@ -1,7 +1,7 @@
 import HeaderAuthPanel from '@/components/auth/HeaderAuthPanel'
 import { auth } from '@/libs/auth'
 import { convertToLetter, fetcher, fetcherWithAuth } from '@/libs/utils'
-import codedangLogo from '@/public/logos/codedang-editor.svg'
+import codedangLogo from '@/public/codedang-editor.svg'
 import type { Contest, ContestProblem, ProblemDetail } from '@/types/type'
 import type { Route } from 'next'
 import Image from 'next/image'
